fix(ci): T1-E4 — enforce engine/vector/graph boundary
Per the implementation plan, engine must not depend on vector/graph
even as dev-deps. This eliminates the coupling by:

- Move 5 cross-crate tests from engine to tests/integration/recovery_cross_crate.rs
- Remove strata-vector/strata-graph dev-deps from crates/engine/Cargo.toml
- Fix CI guard to enforce no deps at all (not just no normal deps)
- Add test_db_open helper to tests/common/mod.rs

Tests moved:
- test_vector_collections_isolated_across_spaces_after_restart
- test_vector_create_collection_registers_space_at_engine_layer
- test_vector_insert_registers_space_at_engine_layer
- test_graph_create_registers_space_at_engine_layer
- test_graph_bulk_insert_empty_input_does_not_register_space
