+++ crates/engine/Cargo.toml
@@ -48,8 +48,6 @@ tempfile = { workspace = true }
 criterion = { workspace = true }
 serde_json = { workspace = true }
 tracing-subscriber = { workspace = true }
-strata-vector = { path = "../vector" }
-strata-graph = { path = "../graph" }
 
 [[bench]]
 name = "transaction_benchmarks"
