+++ crates/engine/tests/recovery_tests.rs
@@ -1212,358 +1212,6 @@ fn test_json_slow_path_skips_secondary_index_storage() {
 }
 
 // =============================================================================
-// Phase 2 — Vector backend isolation by space
-// =============================================================================
-
-/// Phase 2 invariant: two vector collections sharing a name across two
-/// different spaces on the same branch must survive a full DB restart with
-/// independent contents, independent disk caches, and independent deletions.
-///
-/// This test exercises every Phase 2 fix in a single round-trip:
-/// - Closing the DB triggers `freeze_to_disk` (Part 1) — without the
-///   space-aware path helpers, the second tenant clobbers the first
-///   tenant's `.vec` cache file.
-/// - Reopening triggers vector recovery (Part 2) — without enumerating
-///   all spaces via `SpaceIndex::list`, the non-default user-space
-///   collections are silently skipped at startup and only resurface via
-///   the lazy reload fallback.
-/// - Step 7 walks the on-disk layout to catch path-construction
-///   regressions directly.
-/// - Step 8 catches any deletion path that bypasses the space dimension.
-#[test]
-fn test_vector_collections_isolated_across_spaces_after_restart() {
-    use strata_core::primitives::vector::DistanceMetric;
-    use strata_engine::{DatabaseBuilder, SearchSubsystem, SpaceIndex};
-    use strata_vector::{VectorConfig, VectorStore, VectorSubsystem};
-
-    let temp_dir = TempDir::new().unwrap();
-    let path = temp_dir.path().to_path_buf();
-    let branch_id = BranchId::new();
-
-    // -----------------------------------------------------------------
-    // Session 1 — create two same-name collections in different spaces
-    // -----------------------------------------------------------------
-    {
-        // Use DatabaseBuilder so VectorSubsystem.freeze() runs on drop and
-        // actually exercises the freeze_to_disk path we're testing.
-        let db = DatabaseBuilder::new()
-            .with_subsystem(VectorSubsystem)
-            .with_subsystem(SearchSubsystem)
-            .open(&path)
-            .unwrap();
-
-        // Register the user spaces in metadata. The executor does this
-        // automatically via `ensure_space_registered` on the production
-        // write path; tests that bypass the executor must register
-        // explicitly so vector recovery can enumerate them via the
-        // `SpaceIndex::list`-equivalent scan in `recovery::recover_from_db`.
-        let space_index = SpaceIndex::new(db.clone());
-        space_index.register(branch_id, "tenant_a").unwrap();
-        space_index.register(branch_id, "tenant_b").unwrap();
-
-        let store = VectorStore::new(db.clone());
-
-        let cfg_a = VectorConfig::new(3, DistanceMetric::Cosine).unwrap();
-        let cfg_b = VectorConfig::new(3, DistanceMetric::Cosine).unwrap();
-
-        store
-            .create_collection(branch_id, "tenant_a", "embeddings", cfg_a)
-            .unwrap();
-        store
-            .create_collection(branch_id, "tenant_b", "embeddings", cfg_b)
-            .unwrap();
-
-        // tenant_a: two distinct vectors
-        store
-            .insert(
-                branch_id,
-                "tenant_a",
-                "embeddings",
-                "k1",
-                &[1.0, 0.0, 0.0],
-                None,
-            )
-            .unwrap();
-        store
-            .insert(
-                branch_id,
-                "tenant_a",
-                "embeddings",
-                "k2",
-                &[0.0, 1.0, 0.0],
-                None,
-            )
-            .unwrap();
-
-        // tenant_b: same key as tenant_a but a different vector. If the
-        // backends bleed across spaces, k1 in one tenant would observe
-        // the other tenant's embedding.
-        store
-            .insert(
-                branch_id,
-                "tenant_b",
-                "embeddings",
-                "k1",
-                &[0.0, 0.0, 1.0],
-                None,
-            )
-            .unwrap();
-
-        // Sanity check pre-restart
-        let collections_a = store.list_collections(branch_id, "tenant_a").unwrap();
-        assert_eq!(collections_a.len(), 1);
-        assert_eq!(collections_a[0].count, 2);
-
-        let collections_b = store.list_collections(branch_id, "tenant_b").unwrap();
-        assert_eq!(collections_b.len(), 1);
-        assert_eq!(collections_b[0].count, 1);
-
-        // Drop in explicit order so the LAST Arc<Database> released is `db`,
-        // which is what triggers `Drop for Database` (and the freeze hooks)
-        // synchronously before we exit the block. `space_index` and `store`
-        // both hold their own `Arc<Database>` clones, so they must be
-        // dropped first or `drop(db)` only decrements the refcount without
-        // firing Drop.
-        drop(space_index);
-        drop(store);
-        drop(db);
-    }
-
-    // -----------------------------------------------------------------
-    // Disk-layout check (proves Part 1 freeze path fix)
-    // -----------------------------------------------------------------
-    let branch_hex = format!("{:032x}", u128::from_be_bytes(*branch_id.as_bytes()));
-    let vectors_root = path.join("vectors").join(&branch_hex);
-    let tenant_a_vec = vectors_root.join("tenant_a").join("embeddings.vec");
-    let tenant_b_vec = vectors_root.join("tenant_b").join("embeddings.vec");
-
-    assert!(
-        tenant_a_vec.exists(),
-        "tenant_a embeddings.vec must exist at {:?} after freeze",
-        tenant_a_vec
-    );
-    assert!(
-        tenant_b_vec.exists(),
-        "tenant_b embeddings.vec must exist at {:?} after freeze",
-        tenant_b_vec
-    );
-
-    // The legacy (buggy) freeze path wrote `vectors/{branch_hex}/embeddings.vec`
-    // with no space subdirectory. If that path exists, Part 1 regressed.
-    let legacy_path = vectors_root.join("embeddings.vec");
-    assert!(
-        !legacy_path.exists(),
-        "legacy space-less path {:?} must not exist — freeze must use space-aware helpers",
-        legacy_path
-    );
-
-    // -----------------------------------------------------------------
-    // Session 2 — reopen and verify isolation after recovery
-    // -----------------------------------------------------------------
-    {
-        use strata_vector::{CollectionId, VectorBackendState};
-
-        let db = DatabaseBuilder::new()
-            .with_subsystem(VectorSubsystem)
-            .with_subsystem(SearchSubsystem)
-            .open(&path)
-            .unwrap();
-
-        // Direct check of in-memory state IMMEDIATELY after open, before
-        // any lazy-load path can run. This is the regression check for
-        // Part 2: vector recovery must enumerate every space registered
-        // for the branch, not just `default` and `_system_`. The lazy-load
-        // fallback in `ensure_collection_loaded` would otherwise mask the
-        // bug by populating backends on first access.
-        //
-        // CollectionId equality is structural (branch_id + space + name).
-        let state = db.extension::<VectorBackendState>().unwrap();
-        let cid_a = CollectionId::new(branch_id, "tenant_a", "embeddings");
-        let cid_b = CollectionId::new(branch_id, "tenant_b", "embeddings");
-        assert!(
-            state.backends.contains_key(&cid_a),
-            "tenant_a/embeddings backend must be loaded by recovery, not deferred to lazy load",
-        );
-        assert!(
-            state.backends.contains_key(&cid_b),
-            "tenant_b/embeddings backend must be loaded by recovery, not deferred to lazy load",
-        );
-
-        let store = VectorStore::new(db.clone());
-
-        // Both collections are recovered with their original counts.
-        let collections_a = store.list_collections(branch_id, "tenant_a").unwrap();
-        assert_eq!(
-            collections_a.len(),
-            1,
-            "tenant_a should have 1 collection after recovery"
-        );
-        assert_eq!(
-            collections_a[0].count, 2,
-            "tenant_a collection should have 2 vectors after recovery"
-        );
-
-        let collections_b = store.list_collections(branch_id, "tenant_b").unwrap();
-        assert_eq!(
-            collections_b.len(),
-            1,
-            "tenant_b should have 1 collection after recovery"
-        );
-        assert_eq!(
-            collections_b[0].count, 1,
-            "tenant_b collection should have 1 vector after recovery"
-        );
-
-        // Search tenant_a for its own k1 vector. We assert TOP-hit equals k1
-        // AND its score is the perfect-match value, which catches any bug
-        // where k1's stored vector got swapped with another vector (e.g.,
-        // recovery loading tenant_b's data into tenant_a's k1 slot).
-        // Cosine similarity returns 1.0 for identical unit vectors.
-        let hits_a = store
-            .search(
-                branch_id,
-                "tenant_a",
-                "embeddings",
-                &[1.0, 0.0, 0.0],
-                10,
-                None,
-            )
-            .unwrap();
-        assert!(!hits_a.is_empty(), "tenant_a search returned no hits");
-        assert_eq!(
-            hits_a[0].key,
-            "k1",
-            "tenant_a top hit for [1,0,0] should be its own k1, got {:?}",
-            hits_a.iter().map(|m| &m.key).collect::<Vec<_>>()
-        );
-        assert!(
-            (hits_a[0].score - 1.0).abs() < 1e-5,
-            "tenant_a k1 stored vector should match query exactly (score ~1.0); got score {} — \
-             this would fail if k1's data was overwritten with a different vector during recovery",
-            hits_a[0].score,
-        );
-
-        // Search tenant_b for its own k1 vector — different vector, same key.
-        let hits_b = store
-            .search(
-                branch_id,
-                "tenant_b",
-                "embeddings",
-                &[0.0, 0.0, 1.0],
-                10,
-                None,
-            )
-            .unwrap();
-        assert!(!hits_b.is_empty(), "tenant_b search returned no hits");
-        assert_eq!(
-            hits_b[0].key,
-            "k1",
-            "tenant_b top hit for [0,0,1] should be its own k1, got {:?}",
-            hits_b.iter().map(|m| &m.key).collect::<Vec<_>>()
-        );
-        assert!(
-            (hits_b[0].score - 1.0).abs() < 1e-5,
-            "tenant_b k1 stored vector should match query exactly (score ~1.0); got score {}",
-            hits_b[0].score,
-        );
-
-        // Cross-space cross-check: searching tenant_a for tenant_b's exact
-        // vector should still return only tenant_a's vectors. tenant_a
-        // does NOT contain [0,0,1], so its top score must be strictly less
-        // than 1.0 (a perfect match would mean tenant_b's k1 leaked in).
-        let cross_hits = store
-            .search(
-                branch_id,
-                "tenant_a",
-                "embeddings",
-                &[0.0, 0.0, 1.0],
-                10,
-                None,
-            )
-            .unwrap();
-        assert_eq!(
-            cross_hits.len(),
-            2,
-            "tenant_a should return only its own 2 vectors, got {}",
-            cross_hits.len()
-        );
-        // tenant_a contains only [1,0,0] and [0,1,0]; both are perpendicular
-        // to the query [0,0,1], so cosine similarity should be ~0. A score
-        // close to 1.0 would mean tenant_b's [0,0,1] vector leaked into
-        // tenant_a's backend.
-        for hit in &cross_hits {
-            assert!(
-                (hit.score - 1.0).abs() > 0.5,
-                "tenant_a hit {:?} unexpectedly has perfect-match score {}, \
-                 indicating tenant_b's vector may have leaked into tenant_a",
-                hit.key,
-                hit.score,
-            );
-        }
-
-        drop(store);
-        drop(db);
-    }
-
-    // -----------------------------------------------------------------
-    // Cross-space deletion isolation
-    // -----------------------------------------------------------------
-    {
-        let db = DatabaseBuilder::new()
-            .with_subsystem(VectorSubsystem)
-            .with_subsystem(SearchSubsystem)
-            .open(&path)
-            .unwrap();
-        let store = VectorStore::new(db.clone());
-
-        store
-            .delete_collection(branch_id, "tenant_a", "embeddings")
-            .unwrap();
-
-        // tenant_b must still be intact in this same session.
-        let collections_b = store.list_collections(branch_id, "tenant_b").unwrap();
-        assert_eq!(collections_b.len(), 1);
-        assert_eq!(collections_b[0].count, 1);
-
-        let collections_a = store.list_collections(branch_id, "tenant_a").unwrap();
-        assert_eq!(
-            collections_a.len(),
-            0,
-            "tenant_a should be empty after delete"
-        );
-
-        drop(store);
-        drop(db);
-    }
-
-    // Reopen one more time to confirm the deletion + tenant_b survival
-    // both persist across recovery.
-    {
-        let db = DatabaseBuilder::new()
-            .with_subsystem(VectorSubsystem)
-            .with_subsystem(SearchSubsystem)
-            .open(&path)
-            .unwrap();
-        let store = VectorStore::new(db.clone());
-
-        let collections_a = store.list_collections(branch_id, "tenant_a").unwrap();
-        assert_eq!(
-            collections_a.len(),
-            0,
-            "tenant_a deletion must persist across restart"
-        );
-
-        let collections_b = store.list_collections(branch_id, "tenant_b").unwrap();
-        assert_eq!(
-            collections_b.len(),
-            1,
-            "tenant_b must survive tenant_a's deletion across restart"
-        );
-        assert_eq!(collections_b[0].count, 1);
-    }
-}
-
-// =============================================================================
 // Phase 3 — Lower-layer registration, discovery, repair, startup wiring
 // =============================================================================
 //
@@ -1572,9 +1220,7 @@ fn test_vector_collections_isolated_across_spaces_after_restart() {
 //   Part 1 (engine-layer registration)
 //     test_kv_put_registers_space_at_engine_layer
 //     test_json_create_registers_space_at_engine_layer
-//     test_vector_create_collection_registers_space_at_engine_layer
-//     test_vector_insert_registers_space_at_engine_layer
-//     test_graph_create_registers_space_at_engine_layer
+//     (vector/graph tests moved to tests/integration/recovery_cross_crate.rs)
 //
 //   Part 3 (discovery + repair + union list/exists + startup wiring)
 //     test_discover_used_spaces_finds_data_in_unregistered_space
@@ -1662,105 +1308,7 @@ fn test_json_create_registers_space_at_engine_layer() {
     );
 }
 
-#[test]
-fn test_vector_create_collection_registers_space_at_engine_layer() {
-    use strata_core::primitives::vector::DistanceMetric;
-    use strata_vector::{VectorConfig, VectorStore};
-
-    let (db, _temp, branch_id) = setup();
-    let store = VectorStore::new(db.clone());
-
-    let cfg = VectorConfig::new(3, DistanceMetric::Cosine).unwrap();
-    store
-        .create_collection(branch_id, "tenant_vec", "embeddings", cfg)
-        .unwrap();
-
-    assert!(
-        space_metadata_exists(&db, branch_id, "tenant_vec"),
-        "VectorStore::create_collection must persist a SpaceIndex metadata key"
-    );
-}
-
-#[test]
-fn test_vector_insert_registers_space_at_engine_layer() {
-    use strata_core::primitives::vector::DistanceMetric;
-    use strata_vector::{VectorConfig, VectorStore};
-
-    let (db, _temp, branch_id) = setup();
-    let store = VectorStore::new(db.clone());
-
-    // `insert_inner` is the call site `create_collection` already covers,
-    // so we test it on its own with a second-step insert that exercises
-    // the registration helper inside the upsert transaction.
-    let cfg = VectorConfig::new(3, DistanceMetric::Cosine).unwrap();
-    store
-        .create_collection(branch_id, "tenant_vec_insert", "embeddings", cfg)
-        .unwrap();
-    store
-        .insert(
-            branch_id,
-            "tenant_vec_insert",
-            "embeddings",
-            "k1",
-            &[1.0, 0.0, 0.0],
-            None,
-        )
-        .unwrap();
-
-    assert!(
-        space_metadata_exists(&db, branch_id, "tenant_vec_insert"),
-        "VectorStore::insert (insert_inner) must persist a SpaceIndex metadata key"
-    );
-}
-
-#[test]
-fn test_graph_create_registers_space_at_engine_layer() {
-    use strata_graph::types::NodeData;
-    use strata_graph::GraphStore;
-
-    let (db, _temp, branch_id) = setup();
-    let graph = GraphStore::new(db.clone());
-
-    // Add a node directly — `add_node` flows through `txn.graph_add_node`
-    // which calls `ensure_space_registered_in_txn`. We don't call
-    // `create_graph` first, so this is the strict test of the helper.
-    graph
-        .add_node(
-            branch_id,
-            "tenant_graph",
-            "social",
-            "alice",
-            NodeData::default(),
-        )
-        .unwrap();
-
-    assert!(
-        space_metadata_exists(&db, branch_id, "tenant_graph"),
-        "GraphStore::add_node must persist a SpaceIndex metadata key for its space"
-    );
-}
-
-#[test]
-fn test_graph_bulk_insert_empty_input_does_not_register_space() {
-    use strata_graph::GraphStore;
-
-    let (db, _temp, branch_id) = setup();
-    let graph = GraphStore::new(db.clone());
-
-    // Empty bulk insert must NOT create a phantom space metadata entry.
-    // Pre-fix `bulk_insert` registered the space upfront before checking
-    // whether there was anything to write.
-    let (nodes_inserted, edges_inserted) = graph
-        .bulk_insert(branch_id, "phantom_bulk", "g", &[], &[], None)
-        .unwrap();
-    assert_eq!(nodes_inserted, 0);
-    assert_eq!(edges_inserted, 0);
-
-    assert!(
-        !space_metadata_exists(&db, branch_id, "phantom_bulk"),
-        "bulk_insert with empty inputs must not register a phantom space"
-    );
-}
+// Vector/graph registration tests moved to tests/integration/recovery_cross_crate.rs
 
 // --- Part 3 — discovery / repair / union / startup -----------------------
